add: edit-profile in nav
