+++ src/widgets/aside/consts/nav.tsx
@@ -1,0 +1,27 @@
+import IconHome from "@/shared/ui/icons/icon-home";
+
+const navList = [
+  {
+    name: "Home",
+    icon: <IconHome/>,
+    link: "/home"
+  },
+  {
+    name: "Profile",
+    icon: <IconHome/>,
+    link: "/profile"
+  },
+  {
+    name: "Edit Profile",
+    icon: <IconHome/>,
+    link: "/profile/edit"
+  },
+  {
+    name: "Messages",
+    icon: <IconHome/>,
+    link: "/messages"
+  }
+];
+
+
+export default navList;
